+++ src/main/java/com/example/demospringboot/repository/CustomerRepository.java
@@ -1,5 +1,7 @@
-package com.example.demospringboot.domain;
+package com.example.demospringboot.repository;
 
+import com.example.demospringboot.domain.Customer;
+import com.example.demospringboot.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
@@ -15,5 +17,5 @@ public interface CustomerRepository extends JpaRepository<Customer, Long> {
     /*    @Query(value = "SELECT id, name FROM Orders WHERE name = orderName\n" +
                 "INNER JOIN Customers ON Orders.ID = Customers.Order_fk;", nativeQuery = true)*/
     List<Order> getValues(String orderName);
-    List<Customer> findByOrder(String orderName);
+    List<Customer> findByOrder(Order order);
 }
